+++ pyengine/Utils/__init__.py
@@ -4,3 +4,4 @@ from pyengine.Utils.Vec2 import Vec2
 from pyengine.Utils.Logger import loggers
 from pyengine.Utils.Lang import Lang
 from pyengine.Utils.Config import Config
+from pyengine.Utils.Maths import clamp