+++ tests/UtilsTests.py
@@ -3,6 +3,13 @@ from pyengine.Utils import *
 import os
 
 
+class MathsTests(unittest.TestCase):
+    def test_clamp(self):
+        self.assertEqual(clamp(1, 0, 2), 1)
+        self.assertEqual(clamp(3, 0, 2), 2)
+        self.assertEqual(clamp(-1, 0, 2), 0)
+
+
 class ColorTests(unittest.TestCase):
     def setUp(self):
         self.color = Color()